refactor: 도메인 로직인 Car 클래스를 domain 디렉터리로 이동
