+++ docs/REQUIREMENTS.md
@@ -17,8 +17,8 @@
 
 ## 🎯 리팩터링 요구 사항
 
-- [] 도메인 로직은 domain/ 하위로, UI 관련 로직은 view/ 하위에서 관리한다.
-- [] domain/ 하위의 모듈은 view/ 하위의 모듈을 의존하지 않아야 한다.
+- [x] 도메인 로직은 domain/ 하위로, UI 관련 로직은 view/ 하위에서 관리한다.
+- [x] domain/ 하위의 모듈은 view/ 하위의 모듈을 의존하지 않아야 한다.
 - [] 도메인 로직 내에서 테스트하기 어려운 부분을 분리하고, 테스트 가능한 부분에 대해서만 단위 테스트를 작성한다.
 - [] 테스트 코드에서 jest.fn()을 사용하지 않는다.
